python-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==2.3.0
pymupdf==1.23.8
httpx==0.25.2
google-generativeai==0.3.2
anthropic==0.7.8
openai==1.6.1
pydantic==2.5.2
